Never set DEFAULT_SOCKET_DIR from TMPDIR, TEMP or TMP
The temporary directory that is being used to compile dbus is
orthogonal to the directory that we want to use for any
filesystem-backed sockets we might create during runtime. If a user
has chosen to use TMPDIR=/var/tmp or TMPDIR=/home/user/tmp for the
build process, it seems incorrect to hard-code that into libdbus.

We already offer --with-session-socket-dir and --with-test-socket-dir
if a distributor wants to customize these.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>