spec: Document the initial Containers1 interface
Signed-off-by: Simon McVittie <smcv@collabora.com>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=101354

---

Adjust wording as Philip suggested.

Instead of talking about "active" container instances and never
defining what that means, talk about "valid" container instances and
define that term in AddServer(). The general operating principle
is that a container instance object path is valid as long as it
either has connections, or could have connections in future (is
listening for connections).

When implementing Bug #101898 I plan to make the path returned from
AddServer() be considered to be valid immediately, even if the
ServerSocketReadyNotifier has not yet gone off: the important thing
is whether the container instance could still produce connections
later, not whether it is listening for connections right now.