test/containers: Check that containers can't make new containers
We should prevent containers from trying to putting a container in our
container so we can contain while we contain; the implementation doesn't
actually have any concept of nesting or layering, so that would potentially
be privilege escalation.

At the moment, this is just prevented by METHOD_FLAG_PRIVILEGED. If we
remove that flag when we've introduced better resource limits, then
we can specifically restrict this method to not be called by
containers instead: this test will make sure we do.

Signed-off-by: Simon McVittie <smcv@collabora.com>