Remove deprecated support for pam_console/pam_foreground flag-files
This was deprecated in 1.11.18.

pam_console appears to have been in Fedora and Gentoo until 2007.
pam_foreground seems to be specific to Debian and Ubuntu, where it was
unmaintained since 2008 and removed in 2010. The replacement for both
was ConsoleKit, which has itself been superseded by systemd-logind and
ConsoleKit2.

This is all here to support <policy at_console="true">, which is
itself a deprecated/discouraged feature. Backends for those policies
still exist for systemd-logind (Linux-specific, select with
--enable-systemd), and for detecting the (single) user at the console
by probing the ownership of /dev/console or some other hard-coded path
(theoretically cross-platform but probably only useful on Solaris,
select with --enable-console-owner-file/--with-console-owner-file).

Signed-off-by: Simon McVittie <smcv@collabora.com>
Bug: https://bugs.freedesktop.org/101629