Officially deprecate <policy at_console="true">
We've been making noises about discouraging this for a while,
but NEWS and dbus-daemon(1) never actually said it was a bad idea.
Let's fix that.

I would like to be able to point to a documentation web page that
explains what at_console does, how that isn't what you think it does,
why it's discouraged, and what you can/should do instead; but if we
wait for someone to have time to write that, then we'll be waiting a
long time.

Signed-off-by: Simon McVittie <smcv@collabora.com>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=39611
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=101629