driver: Reorganise GetNameOwner, ListQueuedOwners for filtering
This converts them to a mostly early-return pattern (via a goto because
some cleanup is needed).

If the desired name is DBUS_SERVICE_DBUS, we can just early-return
without doing the other work.

Similarly, if the desired name is not syntactically valid, we can
early-return without even asking the registry whether someone owns it,
because nobody can own an invalid name.

Add a specific code path for reporting that the name doesn't exist, so
that when we filter, all code paths that should be indistinguishable
(can't see because it doesn't exist, or because policy says we can't
see it) will be using the same code.

Signed-off-by: Simon McVittie <smcv@collabora.com>