containers test: Check that unsolicited replies can't be sent
To do this, we need to send messages that contain a special message
body, since there is no straightforward way to tell which message
is which otherwise.

Strictly speaking, this only needs the filter on the unconfined
connection (that's where we're sending the unsolicited replies),
but adding the same filter to the unconfined observer connection
and to each confined connection will make life easier for subsequent
tests.

Signed-off-by: Simon McVittie <smcv@collabora.com>